Made pixel-display window non-resizable and gave it proper size
Autofitting (which should happen if you don't set window size or
set it to 0,0) did not work with the display window. Not sure if
its something about it not working when content consists of ImDrawList.
So instead I had to set the size manually, with a little bit of eye-balling
for the right and bottom side padding between window and edge and content.